issue #411. split out 7 classes.
